requirements.txt
streamlit
openai
pypdf
scikit-learn
numpy
python-dotenv
sentence-transformers